Commit para hacer pull

+++ Usuarios-Service/src/main/java/com/example/spring/usuarios/controller/UsuariosController.java
@@ -1,11 +1,15 @@
 package com.example.spring.usuarios.controller;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -22,9 +26,11 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import com.example.spring.usuarios.adapter.UsuarioAdapter;
+import com.example.spring.usuarios.controller.error.UsuarioNotFoundException;
 import com.example.spring.usuarios.model.Usuario;
 import com.example.spring.usuarios.response.UsuarioDTO;
 import com.example.spring.usuarios.service.UsuarioService;
+import com.example.spring.usuarios.utilidades.Utilidades;
 
 import jakarta.validation.Valid;
 
@@ -35,45 +41,42 @@ public class UsuariosController {
 
 	@Autowired
 	UsuarioService servicio;
-	
+
 	@Autowired
 	UsuarioAdapter adaptador;
-	
+
 	@GetMapping
 	public List<UsuarioDTO> findAll() {
 		logger.info("----- LISTADO DE USUARIOS (GET) -----");
 		return adaptador.listaADTO(servicio.findAll());
 	}
-	
+
 	@DeleteMapping("/{id}")
-	public ResponseEntity<UsuarioDTO> deleteById(@PathVariable Long id) {
+	public ResponseEntity<Map<String, Object>> deleteById(@PathVariable Long id) {
 
 		logger.info("----- BORRADO DE USUARIO (DELETE) -----");
-		
-		//Optional<Usuario> usuarioOptional = servicio.findById(id);
-		
-//		if (usuarioOptional.isPresent()) {
-//			servicio.deleteById(id);
-//			return ResponseEntity.ok(adaptador.of(usuarioOptional.get()));
-//		} else {
-//			throw new UsuarioNotFoundException();
-//		}
-		
-		servicio.deleteById(id);
-		
-		return null;
+		Optional<Usuario> usuarioOptional = Optional.ofNullable(servicio.findById(id));
+		Map<String, Object> respuesta = new HashMap<>();
+		HttpStatus status = null;
 
+		if (usuarioOptional.isPresent()) {
+			servicio.deleteById(id);
+			status = HttpStatus.OK;
+			respuesta = Utilidades.usuarioEliminadoJson(adaptador.of(usuarioOptional));
+		}
+
+		return new ResponseEntity<>(respuesta, status);
 	}
-	
+
 	@GetMapping("/{id}")
 	public UsuarioDTO findById(@PathVariable Long id) {
-		
+
 		logger.info("----- LISTADO DE USUARIO POR ID (GET)-----");
-		
-		UsuarioDTO usuario= adaptador.of(servicio.findById(id));
+
+		UsuarioDTO usuario = adaptador.of(servicio.findById(id));
 		return usuario;
 	}
-	
+
 	@PostMapping
 	public ResponseEntity<UsuarioDTO> altaUsuario(@Valid @RequestBody Usuario usuario) {
 		logger.info("------ Alta de usuario (POST)");
@@ -84,7 +87,7 @@ public class UsuariosController {
 
 		return ResponseEntity.created(location).body(result);
 	}
-	
+
 	@PutMapping("/{id}")
 	public ResponseEntity<UsuarioDTO> editarUsuario(@PathVariable int id, @Valid @RequestBody Usuario usuario) {
 		logger.info("------ Editar usuario (PUT)");
@@ -95,5 +98,5 @@ public class UsuariosController {
 
 		return ResponseEntity.created(location).body(result);
 	}
-	
+
 }
